update convert tests with refactored version
